Fix missing background image
Parses the configuration file for the baseURL to extract the path of the URL.
If the path is present, the generated html will get patched to prepend the bath to backgroud image URLs.
This is a very unnsatisfactory workaround for https://github.com/google/docsy/issues/350 and https://github.com/gohugoio/hugo/issues/7714.