Make int tests create workflows in kfp-operator-system namespace
This is needed because the workflowtemplates (which the argo workflows
need) are deployed as part of the operator helm chart in
kfp-operator-system, so we get not found errors if we try to create
workflows in argo. This is a problem we will have to deal with once we
run workflows in provider namespaces.