Fetch Run and RunConfig at webhook top level instead of event processor
Currently, we fetch the Run and RunConfiguration (in order to extract the
output artifacts and add them to the RunCompletion event) in the
EventProcessor code. This adds hidden functionality and side effects to the EventProcessor, which should be
simply converting from RunCompletionEventData to RunCompletionEvent.

So in this commit we bubble up the fetching of k8s resources to the
top level of the runcompletion feed code (in the handleEvent method).

Co-authored-by: Toby Pinfold <28760528+TobyPinfold@users.noreply.github.com>