Fix run completion event trigger service not showing per-request logging (#627)
Fix run completion event trigger service not showing per-request logging due to a regression introduced in ba1d2a0821cdefa037d5e4b9ca061101172edfdb.
In that previous change, the top level context (which had the logger) was no longer being passed into the server, which was using it on a per request basis. This is because contexts should be derived on a per request basis, however when deriving the logger from the new per-request context, this broke logging since that new context did not have a logger.

In this change: introduce a unary server interceptor that will still use a new context per quest in line with golang best practices, but this time it will attach the top level logger into the context which restores logging.