Ignore runconfiguration reconcile condition LastTransitionTime (#622)
The run configuration controller will sometimes get stuck in an infinite loop of reconciling a resource. This happens when a `Condition.LastTransitionTime` is being set by `SetSynchronizationState` and then immediately being compared with the in memory resource, which by definition will be different. This triggers a reconcile; in particular `MergeConditionsInto` is checking `LastTransitionTime` changes.

This change will remove `LastTransitionTime` from `MergeConditionsInto`, which means the controller will not reconcile just because that field was set.